Swap mg and eg in internal representation of Score
Swap mg and eg in internal representation of Score
Instrumentation shows that in make_score(mg, eg) calls, the mg value is
zero in 25,9% of the calls while the eg value is zero in 36,8% of the
calls.

Swapping the internal fields of mg and eg in the internal
representation of Score allows the compiler to optimize away the shift
in (eg << 16) + mg in more cases, thus resulting in a 0.3% speed-up
overall.

No functional change
master (#1)
1 parent 057d710 commit ea41f18e6e6404a6135d98eaeccfbeb177d5a49d
@snicolet snicolet committed